requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-queue
requests
azure-identity
azure-storage-blob
pandas
pyspark
azure-keyvault-secrets
setuptools
wheel
delta
delta-spark
httpx